Rearranged to follow new arch, and finished QA training
