+++ happytransformer/happytasks/happy_mwp/default_args_mwp.py
@@ -1,0 +1,19 @@
+ARGS_MWP_TRAIN = {
+    'max_length': 300,
+    'batch_size': 16,
+    'learning_rate': 5e-5,
+    'epochs': 2,
+
+
+}
+
+ARGS_MWP_TEST = {
+    # eventually we'll add settings
+
+}
+
+ARGS_MWP_EVAL = {
+
+    # eventually we'll add settings
+
+}
