+++ happytransformer/happytasks/happy_qa/default_args_qa.py
@@ -1,0 +1,17 @@
+ARGS_QA_TRAIN = {
+    'max_length': 300,
+    'batch_size': 16,
+    'learning_rate': 5e-5,
+    'epochs': 2,
+
+
+}
+
+ARGS_QA_TEST = {
+    # eventually we'll add settings
+
+}
+
+ARGS_QA_EVAL = {
+    # eventually we'll add settings
+}
