--- happytransformer/trainers/default_args/default_args_qa.py
@@ -1,8 +1,0 @@
-ARGS_QA_TRAINING = {
-    'max_length': 300,
-    'batch_size': 16,
-    'learning_rate': 5e-5,
-    'epochs': 2,
-
-
-}
