+++ tests/test_task_util.py
@@ -2,7 +2,7 @@
 Contains tests for functions found within qa_util.py
 """
 
-from happytransformer.runners.runner_util import SumPair, biggest_sums
+from happytransformer.happytasks.tasks_util import SumPair, biggest_sums
 def test_biggest_sums():
     """
     Tests the biggest_sums function
